+++ irs-edc-client/src/test/java/org/eclipse/tractusx/irs/edc/client/EdcSubmodelClientTest.java
@@ -4,7 +4,7 @@
  *       2022: ISTOS GmbH
  *       2022,2023: Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
  *       2022,2023: BOSCH AG
- * Copyright (c) 2021,2022,2023 Contributors to the Eclipse Foundation
+ * Copyright (c) 2021,2024 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -24,7 +24,6 @@
 package org.eclipse.tractusx.irs.edc.client;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 import static org.eclipse.tractusx.irs.edc.client.cache.endpointdatareference.EndpointDataReferenceStatus.TokenStatus;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
@@ -62,6 +61,8 @@ import org.eclipse.tractusx.irs.component.Relationship;
 import org.eclipse.tractusx.irs.component.enums.BomLifecycle;
 import org.eclipse.tractusx.irs.component.enums.Direction;
 import org.eclipse.tractusx.irs.data.StringMapper;
+import org.eclipse.tractusx.irs.edc.client.cache.endpointdatareference.EndpointDataReferenceCacheService;
+import org.eclipse.tractusx.irs.edc.client.cache.endpointdatareference.EndpointDataReferenceStatus;
 import org.eclipse.tractusx.irs.edc.client.exceptions.ContractNegotiationException;
 import org.eclipse.tractusx.irs.edc.client.exceptions.EdcClientException;
 import org.eclipse.tractusx.irs.edc.client.exceptions.TransferProcessException;
@@ -71,8 +72,6 @@ import org.eclipse.tractusx.irs.edc.client.model.NegotiationResponse;
 import org.eclipse.tractusx.irs.edc.client.model.notification.EdcNotification;
 import org.eclipse.tractusx.irs.edc.client.model.notification.EdcNotificationResponse;
 import org.eclipse.tractusx.irs.edc.client.model.notification.NotificationContent;
-import org.eclipse.tractusx.irs.edc.client.cache.endpointdatareference.EndpointDataReferenceCacheService;
-import org.eclipse.tractusx.irs.edc.client.cache.endpointdatareference.EndpointDataReferenceStatus;
 import org.eclipse.tractusx.irs.testing.containers.LocalTestDataConfigurationAware;
 import org.jetbrains.annotations.NotNull;
 import org.junit.jupiter.api.BeforeEach;
@@ -126,7 +125,8 @@ class EdcSubmodelClientTest extends LocalTestDataConfigurationAware {
         config.getSubmodel().setUrnPrefix("/urn");
         config.getSubmodel().setRequestTtl(Duration.ofMinutes(10));
         testee = new EdcSubmodelClientImpl(config, contractNegotiationService, edcDataPlaneClient,
-                endpointDataReferenceStorage, pollingService, retryRegistry, catalogFacade, endpointDataReferenceCacheService);
+                endpointDataReferenceStorage, pollingService, retryRegistry, catalogFacade,
+                endpointDataReferenceCacheService);
     }
 
     @Test
@@ -134,13 +134,15 @@ class EdcSubmodelClientTest extends LocalTestDataConfigurationAware {
         // arrange
         when(catalogFacade.fetchCatalogByFilter(any(), any(), any())).thenReturn(
                 List.of(CatalogItem.builder().itemId("itemId").build()));
-        when(contractNegotiationService.negotiate(any(), any(), eq(new EndpointDataReferenceStatus(null, TokenStatus.REQUIRED_NEW)))).thenReturn(
+        when(contractNegotiationService.negotiate(any(), any(),
+                eq(new EndpointDataReferenceStatus(null, TokenStatus.REQUIRED_NEW)))).thenReturn(
                 NegotiationResponse.builder().contractAgreementId("agreementId").build());
         final EndpointDataReference ref = mock(EndpointDataReference.class);
         endpointDataReferenceStorage.put("agreementId", ref);
         final String singleLevelBomAsBuiltJson = readSingleLevelBomAsBuiltData();
         when(edcDataPlaneClient.getData(eq(ref), any())).thenReturn(singleLevelBomAsBuiltJson);
-        when(endpointDataReferenceCacheService.getEndpointDataReference("assetId")).thenReturn(new EndpointDataReferenceStatus(null, TokenStatus.REQUIRED_NEW));
+        when(endpointDataReferenceCacheService.getEndpointDataReference("assetId")).thenReturn(
+                new EndpointDataReferenceStatus(null, TokenStatus.REQUIRED_NEW));
 
         // act
         final var result = testee.getSubmodelRawPayload(ENDPOINT_ADDRESS, "suffix", "assetId");
@@ -161,7 +163,8 @@ class EdcSubmodelClientTest extends LocalTestDataConfigurationAware {
         final EndpointDataReference ref = mock(EndpointDataReference.class);
         endpointDataReferenceStorage.put("agreementId", ref);
         when(edcDataPlaneClient.sendData(ref, notification)).thenReturn(() -> true);
-        when(endpointDataReferenceCacheService.getEndpointDataReference(any())).thenReturn(new EndpointDataReferenceStatus(null, TokenStatus.REQUIRED_NEW));
+        when(endpointDataReferenceCacheService.getEndpointDataReference(any())).thenReturn(
+                new EndpointDataReferenceStatus(null, TokenStatus.REQUIRED_NEW));
 
         // act
         final var result = testee.sendNotification(CONNECTOR_ENDPOINT, "notify-request-asset", notification);
@@ -188,7 +191,8 @@ class EdcSubmodelClientTest extends LocalTestDataConfigurationAware {
         when(catalogFacade.fetchCatalogByFilter(any(), any(), any())).thenReturn(
                 List.of(CatalogItem.builder().itemId(existingCatenaXId).build()));
         prepareTestdata(existingCatenaXId, "_singleLevelBomAsBuilt");
-        when(endpointDataReferenceCacheService.getEndpointDataReference(any())).thenReturn(new EndpointDataReferenceStatus(null, TokenStatus.REQUIRED_NEW));
+        when(endpointDataReferenceCacheService.getEndpointDataReference(any())).thenReturn(
+                new EndpointDataReferenceStatus(null, TokenStatus.REQUIRED_NEW));
 
         final String submodelResponse = testee.getSubmodelRawPayload("http://localhost/", "/submodel", ASSET_ID)
                                               .get(5, TimeUnit.SECONDS);
@@ -202,7 +206,8 @@ class EdcSubmodelClientTest extends LocalTestDataConfigurationAware {
         when(catalogFacade.fetchCatalogByFilter(any(), any(), any())).thenReturn(
                 List.of(CatalogItem.builder().itemId(catenaXId).build()));
         prepareTestdata(catenaXId, "_singleLevelBomAsPlanned");
-        when(endpointDataReferenceCacheService.getEndpointDataReference(any())).thenReturn(new EndpointDataReferenceStatus(null, TokenStatus.REQUIRED_NEW));
+        when(endpointDataReferenceCacheService.getEndpointDataReference(any())).thenReturn(
+                new EndpointDataReferenceStatus(null, TokenStatus.REQUIRED_NEW));
 
         final String submodelResponse = testee.getSubmodelRawPayload("http://localhost/", "/submodel", ASSET_ID)
                                               .get(5, TimeUnit.SECONDS);
@@ -216,7 +221,8 @@ class EdcSubmodelClientTest extends LocalTestDataConfigurationAware {
         when(catalogFacade.fetchCatalogByFilter(any(), any(), any())).thenReturn(
                 List.of(CatalogItem.builder().itemId(catenaXId).build()));
         prepareTestdata(catenaXId, "_singleLevelBomAsSpecified");
-        when(endpointDataReferenceCacheService.getEndpointDataReference(any())).thenReturn(new EndpointDataReferenceStatus(null, TokenStatus.REQUIRED_NEW));
+        when(endpointDataReferenceCacheService.getEndpointDataReference(any())).thenReturn(
+                new EndpointDataReferenceStatus(null, TokenStatus.REQUIRED_NEW));
 
         final String submodelResponse = testee.getSubmodelRawPayload("http://localhost/", "/submodel", ASSET_ID)
                                               .get(5, TimeUnit.SECONDS);
@@ -230,7 +236,8 @@ class EdcSubmodelClientTest extends LocalTestDataConfigurationAware {
         when(catalogFacade.fetchCatalogByFilter(any(), any(), any())).thenReturn(
                 List.of(CatalogItem.builder().itemId(catenaXId).build()));
         prepareTestdata(catenaXId, "_singleLevelUsageAsBuilt");
-        when(endpointDataReferenceCacheService.getEndpointDataReference(any())).thenReturn(new EndpointDataReferenceStatus(null, TokenStatus.REQUIRED_NEW));
+        when(endpointDataReferenceCacheService.getEndpointDataReference(any())).thenReturn(
+                new EndpointDataReferenceStatus(null, TokenStatus.REQUIRED_NEW));
 
         final String submodelResponse = testee.getSubmodelRawPayload("http://localhost/", "/submodel", ASSET_ID)
                                               .get(5, TimeUnit.SECONDS);
@@ -245,8 +252,8 @@ class EdcSubmodelClientTest extends LocalTestDataConfigurationAware {
         when(catalogFacade.fetchCatalogByFilter(any(), any(), any())).thenReturn(
                 List.of(CatalogItem.builder().itemId(catenaXId).build()));
         prepareTestdata(catenaXId, "_singleLevelBomAsBuilt");
-        when(endpointDataReferenceCacheService.getEndpointDataReference(ASSET_ID)).thenReturn(new EndpointDataReferenceStatus(null, TokenStatus.REQUIRED_NEW));
-
+        when(endpointDataReferenceCacheService.getEndpointDataReference(ASSET_ID)).thenReturn(
+                new EndpointDataReferenceStatus(null, TokenStatus.REQUIRED_NEW));
 
         final String submodelResponse = testee.getSubmodelRawPayload("http://localhost/", "/submodel", ASSET_ID)
                                               .get(5, TimeUnit.SECONDS);
@@ -260,7 +267,8 @@ class EdcSubmodelClientTest extends LocalTestDataConfigurationAware {
         when(catalogFacade.fetchCatalogByFilter("https://connector.endpoint.com" + PROVIDER_SUFFIX,
                 "https://w3id.org/edc/v0.0.1/ns/id", ASSET_ID)).thenReturn(createCatalog(ASSET_ID, 3));
         prepareTestdata(existingCatenaXId, "_serialPart");
-        when(endpointDataReferenceCacheService.getEndpointDataReference(ASSET_ID)).thenReturn(new EndpointDataReferenceStatus(null, TokenStatus.REQUIRED_NEW));
+        when(endpointDataReferenceCacheService.getEndpointDataReference(ASSET_ID)).thenReturn(
+                new EndpointDataReferenceStatus(null, TokenStatus.REQUIRED_NEW));
 
         final String submodelResponse = testee.getSubmodelRawPayload("https://connector.endpoint.com",
                 "/shells/{aasIdentifier}/submodels/{submodelIdentifier}/submodel", ASSET_ID).get(5, TimeUnit.SECONDS);
@@ -276,7 +284,8 @@ class EdcSubmodelClientTest extends LocalTestDataConfigurationAware {
         final String target = URLEncoder.encode(ASSET_ID, StandardCharsets.UTF_8);
         when(catalogFacade.fetchCatalogByFilter("https://connector.endpoint.com" + PROVIDER_SUFFIX,
                 "https://w3id.org/edc/v0.0.1/ns/id", ASSET_ID)).thenReturn(createCatalog(target, 3));
-        when(endpointDataReferenceCacheService.getEndpointDataReference(ASSET_ID)).thenReturn(new EndpointDataReferenceStatus(null, TokenStatus.REQUIRED_NEW));
+        when(endpointDataReferenceCacheService.getEndpointDataReference(ASSET_ID)).thenReturn(
+                new EndpointDataReferenceStatus(null, TokenStatus.REQUIRED_NEW));
 
         final String submodelResponse = testee.getSubmodelRawPayload("https://connector.endpoint.com",
                 "/shells/{aasIdentifier}/submodels/{submodelIdentifier}/submodel", ASSET_ID).get(5, TimeUnit.SECONDS);
@@ -292,7 +301,8 @@ class EdcSubmodelClientTest extends LocalTestDataConfigurationAware {
         when(catalogFacade.fetchCatalogByFilter(any(), any(), any())).thenReturn(
                 List.of(CatalogItem.builder().itemId(parentCatenaXId).build()));
         prepareTestdata(parentCatenaXId, "_singleLevelBomAsBuilt");
-        when(endpointDataReferenceCacheService.getEndpointDataReference(ASSET_ID)).thenReturn(new EndpointDataReferenceStatus(null, TokenStatus.REQUIRED_NEW));
+        when(endpointDataReferenceCacheService.getEndpointDataReference(ASSET_ID)).thenReturn(
+                new EndpointDataReferenceStatus(null, TokenStatus.REQUIRED_NEW));
         final String relationshipsJson = testee.getSubmodelRawPayload("http://localhost/", "_singleLevelBomAsBuilt",
                 ASSET_ID).get(5, TimeUnit.SECONDS);
 
@@ -326,8 +336,9 @@ class EdcSubmodelClientTest extends LocalTestDataConfigurationAware {
         final String agreementId = "agreementId";
         when(catalogFacade.fetchCatalogByFilter(any(), any(), any())).thenReturn(
                 List.of(CatalogItem.builder().itemId("asset-id").build()));
-        when(contractNegotiationService.negotiate(any(), any(), eq(new EndpointDataReferenceStatus(null, TokenStatus.REQUIRED_NEW))))
-                .thenReturn(NegotiationResponse.builder().contractAgreementId(agreementId).build());
+        when(contractNegotiationService.negotiate(any(), any(),
+                eq(new EndpointDataReferenceStatus(null, TokenStatus.REQUIRED_NEW)))).thenReturn(
+                NegotiationResponse.builder().contractAgreementId(agreementId).build());
         final EndpointDataReference expected = mock(EndpointDataReference.class);
         endpointDataReferenceStorage.put(agreementId, expected);
 
@@ -348,8 +359,9 @@ class EdcSubmodelClientTest extends LocalTestDataConfigurationAware {
         final String agreementId = "agreementId";
         when(catalogFacade.fetchCatalogByFilter(any(), any(), any())).thenReturn(
                 List.of(CatalogItem.builder().itemId("asset-id").build()));
-        when(contractNegotiationService.negotiate(any(), any(), eq(new EndpointDataReferenceStatus(null, TokenStatus.REQUIRED_NEW))))
-                .thenReturn(NegotiationResponse.builder().contractAgreementId(agreementId).build());
+        when(contractNegotiationService.negotiate(any(), any(),
+                eq(new EndpointDataReferenceStatus(null, TokenStatus.REQUIRED_NEW)))).thenReturn(
+                NegotiationResponse.builder().contractAgreementId(agreementId).build());
         final EndpointDataReference expected = mock(EndpointDataReference.class);
         endpointDataReferenceStorage.put(agreementId, expected);
 
@@ -362,17 +374,18 @@ class EdcSubmodelClientTest extends LocalTestDataConfigurationAware {
     }
 
     @Test
-    void shouldUseCachedEndpointReferenceValueWhenTokenIsValid() throws EdcClientException, ExecutionException, InterruptedException {
+    void shouldUseCachedEndpointReferenceValueWhenTokenIsValid()
+            throws EdcClientException, ExecutionException, InterruptedException {
         // given
         when(endpointDataReferenceCacheService.getEndpointDataReference(any())).thenReturn(
-                new EndpointDataReferenceStatus(EndpointDataReference.Builder.newInstance().endpoint("").authKey("").authCode("").build(),
+                new EndpointDataReferenceStatus(
+                        EndpointDataReference.Builder.newInstance().endpoint("").authKey("").authCode("").build(),
                         TokenStatus.VALID));
         final String value = "result";
         when(edcDataPlaneClient.getData(any(), any())).thenReturn(value);
 
         // when
-        final var resultFuture
-                = testee.getSubmodelRawPayload(ENDPOINT_ADDRESS, "suffix", "assetId");
+        final var resultFuture = testee.getSubmodelRawPayload(ENDPOINT_ADDRESS, "suffix", "assetId");
 
         // then
         final String result = resultFuture.get();
@@ -381,18 +394,17 @@ class EdcSubmodelClientTest extends LocalTestDataConfigurationAware {
     }
 
     @Test
-    void shouldCreateCacheRecordWhenTokenIsNotValid()
-            throws EdcClientException {
+    void shouldCreateCacheRecordWhenTokenIsNotValid() throws EdcClientException {
         // given
         when(catalogFacade.fetchCatalogByFilter(any(), any(), any())).thenReturn(
                 List.of(CatalogItem.builder().itemId("itemId").build()));
-        when(contractNegotiationService.negotiate(any(), any(), eq(new EndpointDataReferenceStatus(null, TokenStatus.REQUIRED_NEW)))).thenReturn(
+        when(contractNegotiationService.negotiate(any(), any(),
+                eq(new EndpointDataReferenceStatus(null, TokenStatus.REQUIRED_NEW)))).thenReturn(
                 NegotiationResponse.builder().contractAgreementId("agreementId").build());
         final EndpointDataReference ref = mock(EndpointDataReference.class);
         endpointDataReferenceStorage.put("agreementId", ref);
         when(endpointDataReferenceCacheService.getEndpointDataReference(any())).thenReturn(
-                new EndpointDataReferenceStatus(null,
-                        TokenStatus.REQUIRED_NEW));
+                new EndpointDataReferenceStatus(null, TokenStatus.REQUIRED_NEW));
 
         // when
         testee.getSubmodelRawPayload(ENDPOINT_ADDRESS, "suffix", "assetId");
@@ -405,7 +417,8 @@ class EdcSubmodelClientTest extends LocalTestDataConfigurationAware {
     private void prepareTestdata(final String catenaXId, final String submodelDataSuffix)
             throws ContractNegotiationException, IOException, UsagePolicyException, TransferProcessException {
 
-        when(contractNegotiationService.negotiate(any(), any(), eq(new EndpointDataReferenceStatus(null, TokenStatus.REQUIRED_NEW)))).thenReturn(
+        when(contractNegotiationService.negotiate(any(), any(),
+                eq(new EndpointDataReferenceStatus(null, TokenStatus.REQUIRED_NEW)))).thenReturn(
                 NegotiationResponse.builder().contractAgreementId("agreementId").build());
         final EndpointDataReference ref = mock(EndpointDataReference.class);
         endpointDataReferenceStorage.put("agreementId", ref);