+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Updated license header to "Copyright (c) 2021,2024 Contributors to the Eclipse Foundation"
+
 ## [4.4.0] - 2024-01-15
 ### Added
 - Added EDR token cache to reuse token after contract negotiation